Extract writeGradlePluginModule
